libgcrypt20 (1.11.0-7) unstable; urgency=medium
.
  [ Simon McVittie ]
  * Add an autopkgtest smoke-test for linking to libgcrypt.
    Superficial tests like this one are a surprisingly effective way to
    detect packaging regressions. Closes: #1088243
.
  [ Andreas Metzler ]
  * Sync with Ubuntu 1.11.0-6ubuntu1:
    Pull upstream patch to fix FTBFS on s390x during build-time tests:
    + lp2083245-disable-sha3-s390x-acceleration-for-cshake.patch
    (LP: #2083245)
  * Run wrap-and-sort -ast.
.
libgcrypt20 (1.11.0-6) unstable; urgency=medium
.
  * 30_mpi-ec-inline-reduce-register-pressure-on-32-bit-ARM.patchn from
    upstream GIT master: Fix FTBFS on arm with gcc-14. Closes: #1077326
.
libgcrypt20 (1.11.0-5) unstable; urgency=medium
.
  * Upload to unstable.
.
libgcrypt20 (1.11.0-4) experimental; urgency=medium
.
  * Add missing Breaks/Replaces to libgcrypt-bin. Closes: #1076625
.
libgcrypt20 (1.11.0-3) experimental; urgency=medium
.
  * Point watchfile to releases instead of snapshots again.
  * Move helper binaries from -dev package to newly added libgcrypt-bin
    package, mark libgcrypt20-dev Multi-Arch: same. Closes: #714589
.
libgcrypt20 (1.11.0-2) unstable; urgency=low
.
  * Upload to unstable.
.
libgcrypt20 (1.11.0-1) experimental; urgency=low
.
  * New upstream version.
    + Drop 30_add_missing_files_from_master.diff.
    + Update libgcrypt20.symbols to refer to release instead of beta version.
.
libgcrypt20 (1.11.0~beta450-1) experimental; urgency=low
.
  * New upstream snapshot.
    + Drop 30_m4-Include-_AM_PATH_GPGRT_CONFIG-definition.patch.
    + Update symbol file. Add new functions. Bump version requirement
      on all symbols. (Many algorithms, enums extended, etc.)
    + 30_add_missing_files_from_master.diff: Add some files missing in tarball
      from upstream GIT master.
    + Update debian/copyright.
    + Version libgpg-error-dev build-dependency.
    + Drop debian/libgcrypt-config.1, libgcrypt-config is not shipped anymore.
.
libgcrypt20 (1.10.3-3) unstable; urgency=medium
.
  * 30_m4-Include-_AM_PATH_GPGRT_CONFIG-definition.patch from upstream GIT
    master: Update libgcrypt.m4 to let AM_PATH_LIBGCRYPT continue to work
    without preceding AM_PATH_GPG_ERROR() when libgcrypt-config is removed.
.
libgcrypt20 (1.10.3-2) unstable; urgency=medium
.
  * Upload to unstable.
.
libgcrypt20 (1.10.3-1) experimental; urgency=medium
.
  * New upstream release.
    + Update symbol file.
.
libgcrypt20 (1.10.2-3) unstable; urgency=medium
.
  [ Simon Josefsson ]
  * Update Homepage: URL.
.
  [ Andreas Metzler ]
  * Drop --insert-timestamp linker option on mingw*, binutils 2.41 should use
    SOURCE_DATE_EPOCH automatically and the Debian package has dropped the
    patch to add the --insert-timestamp option. Closes: #1052219
.
libgcrypt20 (1.10.2-2) unstable; urgency=medium
.
  * Upload to unstable.
.
libgcrypt20 (1.10.2-1) experimental; urgency=medium
.
  [ Helmut Grohne ]
  * Support the noudeb build profile. Closes: #1024943
.
  [ Andreas Metzler ]
  * New upstream release.
    + Drop superfluous patches.

+++ cipher/keccak_permute_64.h
@@ -4,7 +4,7 @@
  * This file is part of Libgcrypt.
  *
  * Libgcrypt is free software; you can redistribute it and/or modify
- * it under the terms of the GNU Lesser general Public License as
+ * it under the terms of the GNU Lesser General Public License as
  * published by the Free Software Foundation; either version 2.1 of
  * the License, or (at your option) any later version.
  *
@@ -292,7 +292,7 @@ KECCAK_F1600_PERMUTE_FUNC_NAME(KECCAK_STATE *hd)
 
 static unsigned int
 KECCAK_F1600_ABSORB_FUNC_NAME(KECCAK_STATE *hd, int pos, const byte *lanes,
-			      unsigned int nlanes, int blocklanes)
+			      size_t nlanes, int blocklanes)
 {
   unsigned int burn = 0;
